+++ meteor-app/imports/api/events/server/publications.js
@@ -1,5 +1,6 @@
 import { Meteor } from 'meteor/meteor'
 import Events from '../schema'
+import '../methods'
 
 Meteor.publish('all.events', () => {
   return Events.find({})
@@ -14,17 +15,3 @@ Meteor.publish('add.events', () => {
 })
 
 Meteor.publish('update.Events')
-
-Meteor.methods({
-  'rm.Events': id => {
-    Events.remove(id)
-  },
-  'update.Events': form => {
-    const id = form._id
-    delete form._id
-    Events.update(id, { $set: form })
-  },
-  'add.Events': form => {
-    Events.insert(form)
-  }
-})