+++ meteor-app/imports/api/settings/server/publications.js
@@ -2,11 +2,10 @@ import { Meteor } from 'meteor/meteor'
 import Settings from '../schema'
 import '../methods'
 
-
 Meteor.publish('all.settings', () => {
   return Settings.find({})
 })
 
 Meteor.publish('id.settings', id => {
-  return [Settings.find(id), ]
+  return [Settings.find(id)]
 })