+++ meteor-app/imports/test/generated-factories.js
@@ -1,0 +1,15 @@
+// Generated code to import all of the generated factories
+
+//
+// This file contains a list of factories for testing
+// It is generated from a list of modules
+//
+import './factory.events'
+import './factory.promos'
+import './factory.products'
+import './factory.sessions'
+import './factory.parts'
+import './factory.service-items'
+import './factory.charges'
+import './factory.settings'
+