+++ meteor-app/imports/ui/generated-routes.js
@@ -17,12 +17,12 @@ import SettingsList from '/imports/ui/settings/lister'
 export default GeneratedRoute = () => (
   <>
     <Route path="/admin/events" component={EventsList} />
-    <Route path="/admin/promos" component={PromosList} />
-    <Route path="/admin/products" component={ProductsList} />
-    <Route path="/admin/sessions" component={SessionsList} />
-    <Route path="/admin/parts" component={PartsList} />
-    <Route path="/admin/service-items" component={ServiceItemsList} />
-    <Route path="/admin/charges" component={ChargesList} />
-    <Route path="/admin/settings" component={SettingsList} />
+<Route path="/admin/promos" component={PromosList} />
+<Route path="/admin/products" component={ProductsList} />
+<Route path="/admin/sessions" component={SessionsList} />
+<Route path="/admin/parts" component={PartsList} />
+<Route path="/admin/service-items" component={ServiceItemsList} />
+<Route path="/admin/charges" component={ChargesList} />
+<Route path="/admin/settings" component={SettingsList} />
   </>
 )