+++ meteor-app/imports/ui/member/member-edit-container.js
@@ -13,7 +13,7 @@ const error = new ReactiveVar(false)
 const msg = new ReactiveVar('')
 const newId = new ReactiveVar('')
 
-export default withTracker(props => {
+export default withTracker((props) => {
   // need to do something smarter than this...
   function setError(e) {
     newId.set(null)
@@ -31,12 +31,16 @@ export default withTracker(props => {
     Alert.success(message)
   }
 
-  const setMember = async formData => {
+  const setMember = async (formData) => {
     if (props.member != null) {
       // we are updating the member
       debug('updating member', formData)
       try {
-        const res = await Meteor.callAsync('members.update', formData._id, formData)
+        const res = await Meteor.callAsync(
+          'members.update',
+          formData._id,
+          formData
+        )
         setSuccess('Member Saved', formData._id)
         return res
       } catch (e) {
@@ -48,9 +52,13 @@ export default withTracker(props => {
 
   const updateMemberPassword = async (formData, confirmPass) => {
     if (props.member != null) {
-      debug('updating member password', formData)
+      // debug('updating member password', formData)
       try {
-        const res = await Meteor.callAsync('updateMemberPassword', formData, confirmPass)
+        const res = await Meteor.callAsync(
+          'updateMemberPassword',
+          formData,
+          confirmPass
+        )
         setSuccess('Member password saved', formData.userId)
         return res
       } catch (e) {
@@ -65,6 +73,7 @@ export default withTracker(props => {
   return {
     setMember,
     updateMemberPassword,
+    setMember,
     error: error.get(),
     success: success.get(),
     message: msg.get(),
@@ -72,6 +81,6 @@ export default withTracker(props => {
     newId: newId.get(),
     resetId: () => newId.set(''),
     member: props.member ? props.member : null,
-    schemas: getSchemas(`${Meteor.settings.public.recruit}Edit`)
+    schemas: getSchemas(`${Meteor.settings.public.recruit}Edit`),
   }
 })(MemberEditForm)