+++ meteor-app/imports/ui/service/customer.js
@@ -3,7 +3,14 @@ import { ServiceContext } from './service-context'
 import { Formik, Form, useField } from 'formik'
 import * as Yup from 'yup'
 import moment from 'moment'
-import { Header, Segment, Grid, Input, Button, Checkbox } from 'semantic-ui-react'
+import {
+  Header,
+  Segment,
+  Grid,
+  Input,
+  Button,
+  Checkbox,
+} from 'semantic-ui-react'
 import DatePicker from 'react-datepicker'
 import 'react-datepicker/dist/react-datepicker.css'
 
@@ -12,7 +19,10 @@ const Client = () => {
   const initialState = { ...state }
 
   const validationSchema = Yup.object().shape({
-    name: Yup.string().max(49, 'Name must be less than 50 characters'),
+    name: Yup.string().max(
+      49,
+      'Name must be less than 50 characters'
+    ),
     email: Yup.string().email('Email must be a valid email format'),
     phone: Yup.string()
       .min(3, 'Phone number must be greater than 2 characters')
@@ -21,7 +31,9 @@ const Client = () => {
     model: Yup.string(),
     color: Yup.string().required('Required'),
     bikeValue: Yup.number().required('Required'),
-    assessor: Yup.string().required('Required').max(49, 'Name must be less than 50 characters'),
+    assessor: Yup.string()
+      .required('Required')
+      .max(49, 'Name must be less than 50 characters'),
   })
 
   const TextInput = ({ label, ...props }) => {
@@ -29,7 +41,13 @@ const Client = () => {
     return (
       <div style={{ marginTop: '10px' }}>
         <b>{label}</b>
-        <Input className="text-input" placeholder={label} {...field} {...props} fluid />
+        <Input
+          className="text-input"
+          placeholder={label}
+          {...field}
+          {...props}
+          fluid
+        />
         {meta.touched && meta.error ? (
           <div className="error" style={{ color: 'red' }}>
             {meta.error}
@@ -58,17 +76,38 @@ const Client = () => {
               <Segment>
                 <Header content="Customer Details" dividing />
                 <TextInput id="name-input" label="Name" name="name" />
-                <TextInput id="email-input" label="Email" name="email" />
-                <TextInput id="phone-input" label="Phone" name="phone" />
+                <TextInput
+                  id="email-input"
+                  label="Email"
+                  name="email"
+                />
+                <TextInput
+                  id="phone-input"
+                  label="Phone"
+                  name="phone"
+                />
               </Segment>
             </Grid.Column>
             <Grid.Column>
               <Segment>
                 <Header content="Bike Details" dividing />
                 <TextInput id="make-input" label="Make" name="make" />
-                <TextInput id="model-input" label="Model" name="model" />
-                <TextInput id="colour-input" label="Colour" name="color" />
-                <TextInput id="value-input" label="Value" name="bikeValue" type="number" />
+                <TextInput
+                  id="model-input"
+                  label="Model"
+                  name="model"
+                />
+                <TextInput
+                  id="colour-input"
+                  label="Colour"
+                  name="color"
+                />
+                <TextInput
+                  id="value-input"
+                  label="Value"
+                  name="bikeValue"
+                  type="number"
+                />
               </Segment>
             </Grid.Column>
             <Grid.Column>
@@ -84,14 +123,23 @@ const Client = () => {
                     setFieldValue('pickupDate', date)
                   }}
                 />
-                <TextInput id="assessor" label="Assessor" name="assessor" />
+                <TextInput
+                  id="assessor"
+                  label="Assessor"
+                  name="assessor"
+                />
                 <br />
                 <Checkbox
                   id="temp-bike-checkbox"
                   defaultChecked={state.temporaryBike}
                   label="Temporary Bike"
                   name="temporaryBike"
-                  onChange={() => setFieldValue('temporaryBike', !values.temporaryBike)}
+                  onChange={() =>
+                    setFieldValue(
+                      'temporaryBike',
+                      !values.temporaryBike
+                    )
+                  }
                 />
                 <br />
                 <Checkbox
@@ -99,7 +147,9 @@ const Client = () => {
                   defaultChecked={state.urgent}
                   label="Urgent"
                   name="urgent"
-                  onChange={() => setFieldValue('urgent', !values.urgent)}
+                  onChange={() =>
+                    setFieldValue('urgent', !values.urgent)
+                  }
                 />
                 <br />
                 <Checkbox
@@ -107,7 +157,9 @@ const Client = () => {
                   defaultChecked={state.sentimental}
                   label="Sentimental"
                   name="sentimental"
-                  onChange={() => setFieldValue('sentimental', !values.sentimental)}
+                  onChange={() =>
+                    setFieldValue('sentimental', !values.sentimental)
+                  }
                 />
                 <br />
                 <Checkbox
@@ -115,7 +167,9 @@ const Client = () => {
                   defaultChecked={state.isRefurbish}
                   label="Refurbished"
                   name="isRefurbish"
-                  onChange={() => setFieldValue('isRefurbish', !values.isRefurbish)}
+                  onChange={() =>
+                    setFieldValue('isRefurbish', !values.isRefurbish)
+                  }
                 />
                 <br />
                 <Checkbox
@@ -127,7 +181,12 @@ const Client = () => {
               </Segment>
             </Grid.Column>
           </Grid>
-          <Button type="submit" id="service-form-button" content="Submit" color="blue" />
+          <Button
+            type="submit"
+            id="service-form-button"
+            content="Submit"
+            color="blue"
+          />
         </Form>
       )}
     </Formik>