+++ meteor-app/tests/cypress/integration/b2b/e2e-add-volunteer.js
@@ -9,9 +9,8 @@ Cypress.on('uncaught:exception', (err, runnable) => {
 
 describe('Create member', function() {
   it('Open form - about you', function() {
-    rmEddie('Eddie Mercx')
     cy.visit('/kiosk')
-
+    rmEddie('Eddie Mercx')
     cy.get('button')
       .contains('Register')
       .click()