+++ templates/imports/test/generated-factories.js
@@ -1,0 +1,8 @@
+// Generated code to import all of the generated factories
+
+//
+// This file contains a list of factories for testing
+// It is generated from a list of modules
+//
+<%= factoryImports %>
+