+++ meteor-app/CHANGELOG.md
@@ -1,11 +1,24 @@
 # Changelog
 
 ## Next
+
 - B2B generic info and actionemail templates
 - Send receipt email
 - Import charges from pinpayments
 - Added view of charges with 'refresh' button
 - Added the ability to adjust name or prices on service tags for custom jobs.
+- Add reconciliation to charges and carts, improved layouts
+- Add admin button to edit user details
+- Add b2b member portal
+- Fix coding problems with certain test scripts
+- Add new service test scripts for input and UI
+- Refactor factories code
+- Added member.paymentEmails array
+- Fix email to be lower case
+- Remove settings symlinks, adjust package.json
+- Fix the event import
+- Fix SMS sending to send message as typed
+- Fix SMS sending to detect price change and update DB to suit
 
 ## 1.2.4